module-2-nemo-agent-toolkit/nat_workflows/simple_llm_call/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[project]
name = "simple_llm_call"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[langchain]",
]
requires-python = ">=3.12"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
simple_llm_call = "simple_llm_call.register"